+++ src/main/java/com/softwareverde/bitbalancer/configuration/ConfigurationParser.java
@@ -36,7 +36,7 @@ public class ConfigurationParser {
             return null;
         }
 
-        final Long maxOrphanDepth = json.getLong("maxOrphanDepth");
+        final Integer maxOrphanDepth = json.getInteger("maxOrphanDepth");
 
         final MutableList<NodeProperties> nodePropertiesList = new MutableList<>();
         for (int i = 0; i < nodesJson.length(); ++i) {
