+++ src/main/java/com/softwareverde/bitbalancer/proxy/RpcProxyServer.java
@@ -168,7 +168,7 @@ public class RpcProxyServer {
         bitcoinRpcConnector.subscribeToNotifications(callback);
     }
 
-    public RpcProxyServer(final Integer port, final List<RpcConfiguration> rpcConfigurations, final ZmqConfiguration zmqConfiguration, final Long blockTemplateCacheDuration, final Long maxOrphanDepth) {
+    public RpcProxyServer(final Integer port, final List<RpcConfiguration> rpcConfigurations, final ZmqConfiguration zmqConfiguration, final Long blockTemplateCacheDuration, final Integer maxOrphanDepth) {
         _port = port;
         _rpcConfigurations = rpcConfigurations.asConst();
         _nodeSelector = new HashMapNodeSelector(_rpcConfigurations, maxOrphanDepth);
